Phase 5: close the two limits phase 4 left
5a -- the structure map kept only the first entity of a shared name, so a
class two files define was simply absent from it. FlowIds moves out of
build_flow into a pure core/ids.py (SharedNames), and build_wiki uses the
same rule the flow map does: qualify a name only when two files define it,
and resolve a base, decorator or import to the referencing file's own
definition or leave it as plain text. On the corpus: 5 names qualified, 57
nodes = 57 vault pages, structure edges 3 -> 7. sample_src byte-identical.
Regression case r20; check_graph c07 now reads a structure entity's bare
name.

5b -- below ~1270px the toolbar clipped. The three toggles are now one
element that fitToolbar() re-parents into the overflow menu whenever the
full row does not fit, and back when it does; the rail cap counts only the
compact row. Measured: compact and fitting at 1000px, toggles working from
inside the menu; the floor is now ~987px.

Found and fixed on the way: moving FlowIds out also deleted _file_of and
_claim, crashing every flow build until the verification chain caught it;
and r20 exposed build_graph crashing when the vault is on another drive
than data/ (os.path.relpath raises on Windows) -- it now falls back to the
absolute path.

+++ tools/check_graph.py
@@ -200,7 +200,9 @@ def c07_name_at_source(c):
         else:
             if n.get("kind") == "module":
                 continue                                 # a synthetic "<File>Module" entity
-            body, want, where = "\n".join(lines), [n["id"]], path
+            # A shared name is file-qualified (`widgets.WidgetStore`); the file spells the
+            # bare name, which is the id's last segment.
+            body, want, where = "\n".join(lines), [short(n["id"])], path
         if want and not any(w in body for w in want):
             out.append(f"{n['id']}: {want[0]!r} does not occur in {where} -- the range points elsewhere")
     return out
